fixed issues and code cleanup based on comments
